+++ CHANGELOG.md
@@ -6,7 +6,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
+- Optional key to quickly switch keyboard language ([#62])
 - Added apostrophe as a pop-up character on the dot key ([#356])
+
+### Changed
+- The space bar now shows the currently selected language ([#62])
+- Space bar cursor control now activates immediately on swipe ([#129])
 
 ## [1.7.0] - 2025-12-16
 ### Added
@@ -116,9 +121,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Initial release
 
 [#47]: https://github.com/FossifyOrg/Keyboard/issues/47
+[#62]: https://github.com/FossifyOrg/Keyboard/issues/62
 [#78]: https://github.com/FossifyOrg/Keyboard/issues/78
 [#100]: https://github.com/FossifyOrg/Keyboard/issues/100
 [#117]: https://github.com/FossifyOrg/Keyboard/issues/117
+[#129]: https://github.com/FossifyOrg/Keyboard/issues/129
 [#133]: https://github.com/FossifyOrg/Keyboard/issues/133
 [#134]: https://github.com/FossifyOrg/Keyboard/issues/134
 [#136]: https://github.com/FossifyOrg/Keyboard/issues/136